separate detail render from detail.py
